+++ src/internal/dbutil/db.go
@@ -9,6 +9,7 @@ import (
 	_ "github.com/jackc/pgx/v4/stdlib"
 	"github.com/jmoiron/sqlx"
 	"github.com/pachyderm/pachyderm/v2/src/internal/backoff"
+	"github.com/pachyderm/pachyderm/v2/src/internal/pachsql"
 	"github.com/sirupsen/logrus"
 )
 
@@ -97,7 +98,7 @@ func GetDSN(opts ...Option) string {
 }
 
 // NewDB creates a new DB.
-func NewDB(opts ...Option) (*sqlx.DB, error) {
+func NewDB(opts ...Option) (*pachsql.DB, error) {
 	dbc := newConfig(opts...)
 	if dbc.name == "" {
 		panic("must specify database name")
@@ -125,7 +126,7 @@ func NewDB(opts ...Option) (*sqlx.DB, error) {
 
 // WaitUntilReady attempts to ping the database until the context is cancelled.
 // Progress information is written to log
-func WaitUntilReady(ctx context.Context, log *logrus.Logger, db *sqlx.DB) error {
+func WaitUntilReady(ctx context.Context, log *logrus.Logger, db *pachsql.DB) error {
 	const period = time.Second
 	const timeout = time.Second
 	log.Infof("waiting for db to be ready...")
